Refactor deck actions with loading state
The deck update and delete actions have been refactored to include a loading state, preventing simultaneous changes. In the redux thunk actions, a Promise wrapper was added to allow the use of async/await syntax. The DeckItem component now also includes a useState hook which handles the loading state during dispatch of actions.